Break up integration tests by endpoint
