+++ api_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/EFForg/starttls-check/checker"
+	"github.com/EFForg/starttls-scanner/db"
 )
 
 func TestPolicyCheck(t *testing.T) {
@@ -16,5 +17,26 @@ func TestPolicyCheck(t *testing.T) {
 	result = api.policyCheck("failmail.com")
 	if result.Status != checker.Failure {
 		t.Errorf("Check should have failed.")
+	}
+}
+
+func TestPolicyCheckWithQueuedDomain(t *testing.T) {
+	defer Teardown()
+
+	domainData := db.DomainData{
+		Name:  "example.com",
+		Email: "postmaster@example.com",
+		State: db.StateUnvalidated,
+	}
+	api.Database.PutDomain(domainData)
+	result := api.policyCheck("example.com")
+	if result.Status != checker.Warning {
+		t.Errorf("Check should have warned.")
+	}
+	domainData.State = db.StateQueued
+	api.Database.PutDomain(domainData)
+	result = api.policyCheck("example.com")
+	if result.Status != checker.Warning {
+		t.Errorf("Check should have warned.")
 	}
 }
